refactor(db): separate database tables from schemas
