+++ backend/app/api/deps.py
@@ -12,7 +12,8 @@ from sqlmodel import Session, select
 from app.core import security
 from app.core.config import settings
 from app.core.db import engine
-from app.models import Attendance, Event, TokenPayload, User
+from app.db import Attendance, Event, User
+from app.schemas import TokenPayload
 
 reusable_oauth2 = OAuth2PasswordBearer(
     tokenUrl=f"{settings.API_V1_STR}/login/access-token"
