+++ backend/app/api/routes/private.py
@@ -5,10 +5,8 @@ from pydantic import BaseModel
 
 from app.api.deps import SessionDep
 from app.core.security import get_password_hash
-from app.models import (
-    User,
-    UserPublic,
-)
+from app.db import User
+from app.schemas import UserPublic
 
 router = APIRouter(tags=["private"], prefix="/private")
