+++ backend/app/db/__init__.py
@@ -1,0 +1,23 @@
+from .tables import (
+    Attendance,
+    Event,
+    EventMealOption,
+    Item,
+    Meal,
+    MealChoice,
+    MealType,
+    PackingItem,
+    User,
+)
+
+__all__ = [
+    "User",
+    "Item",
+    "Event",
+    "Meal",
+    "EventMealOption",
+    "PackingItem",
+    "MealChoice",
+    "Attendance",
+    "MealType",
+]
