+++ backend/app/schemas/__init__.py
@@ -1,0 +1,99 @@
+from .attendance import (
+    MealChoiceCreate,
+    MealChoiceCreateBase,
+    MealChoiceUpdate,
+)
+from .auth import (
+    Message,
+    NewPassword,
+    Token,
+    TokenPayload,
+    UpdatePassword,
+)
+from .event import (
+    EventBase,
+    EventCreate,
+    EventPublic,
+    EventsPublic,
+    EventUpdate,
+)
+from .event_meal_option import (
+    EventMealOptionCreate,
+)
+from .item import (
+    ItemBase,
+    ItemCreate,
+    ItemPublic,
+    ItemsPublic,
+    ItemUpdate,
+)
+from .meal import (
+    MealBase,
+    MealCreate,
+    MealPublic,
+    MealUpdate,
+)
+from .packing import (
+    EventPackingList,
+    PackingItemBase,
+    PackingItemCreate,
+    PackingItemPublic,
+    PackingItemsPublic,
+    PackingItemUpdate,
+)
+from .user import (
+    UserBase,
+    UserCreate,
+    UserPublic,
+    UserRegister,
+    UsersPublic,
+    UserUpdate,
+    UserUpdateMe,
+)
+
+__all__ = [
+    # User schemas
+    "UserBase",
+    "UserCreate",
+    "UserRegister",
+    "UserUpdate",
+    "UserUpdateMe",
+    "UserPublic",
+    "UsersPublic",
+    # Auth schemas
+    "Token",
+    "TokenPayload",
+    "NewPassword",
+    "UpdatePassword",
+    "Message",
+    # Item schemas
+    "ItemBase",
+    "ItemCreate",
+    "ItemUpdate",
+    "ItemPublic",
+    "ItemsPublic",
+    # Event schemas
+    "EventBase",
+    "EventCreate",
+    "EventUpdate",
+    "EventPublic",
+    "EventsPublic",
+    # Meal schemas
+    "MealBase",
+    "MealCreate",
+    "MealUpdate",
+    "MealPublic",
+    # Packing schemas
+    "PackingItemBase",
+    "PackingItemCreate",
+    "PackingItemUpdate",
+    "PackingItemPublic",
+    "PackingItemsPublic",
+    "EventPackingList",
+    # Attendance schemas
+    "MealChoiceCreateBase",
+    "MealChoiceCreate",
+    "MealChoiceUpdate",
+    # Event Meal Option schemas
+    "EventMealOptionCreate",
+]
