+++ backend/app/schemas/item.py
@@ -1,0 +1,27 @@
+from uuid import UUID
+
+from sqlmodel import Field, SQLModel
+
+
+class ItemBase(SQLModel):
+    title: str = Field(min_length=1, max_length=255)
+    description: str | None = Field(default=None, max_length=255)
+    category: str | None = Field(default=None, max_length=100)
+
+
+class ItemCreate(ItemBase):
+    pass
+
+
+class ItemUpdate(ItemBase):
+    title: str | None = Field(default=None, min_length=1, max_length=255)  # type: ignore
+
+
+class ItemPublic(ItemBase):
+    id: UUID
+    owner_id: UUID
+
+
+class ItemsPublic(SQLModel):
+    data: list[ItemPublic]
+    count: int
