+++ backend/app/tests/conftest.py
@@ -6,8 +6,7 @@ from sqlmodel import Session, delete, select
 
 from app.core.config import settings
 from app.core.db import engine, init_db
-from app.main import app
-from app.models import (
+from app.db import (
     Attendance,
     Event,
     EventMealOption,
@@ -17,6 +16,7 @@ from app.models import (
     PackingItem,
     User,
 )
+from app.main import app
 from app.tests.utils.user import authentication_token_from_email
 from app.tests.utils.utils import get_superuser_token_headers
