+++ backend/app/tests/crud/test_user.py
@@ -3,7 +3,8 @@ from sqlmodel import Session
 
 from app import crud
 from app.core.security import verify_password
-from app.models import User, UserCreate, UserUpdate
+from app.db import User
+from app.schemas import UserCreate, UserUpdate
 from app.tests.utils.utils import random_email, random_lower_string
